\name{interimTrial}
\alias{interimTrial}
\title{\code{interimTrial} simulates one survival study with interim analyses based
on the patient data and the gene expression data it is supplied
with.  At each interim analyses the error rate (FDR) and the average power
rate (APR) are both calculated and estimated.}
\usage{
  interimTrial(patdat, gendat, l.2, M.1, M.2, alpha,
    powerThreshold, adjustment, standalone = TRUE)
}
\arguments{
  \item{patdat}{the patient data (arrival times and
  survival times) as generated by
  \code{\link{generatePatientData}}}

  \item{gendat}{the gene expression data as generated by
  \code{\link{generateExpressionData}}}

  \item{l.2}{the length of the follow-up period to be
  simulated}

  \item{M.1}{the number of analyses during the recruitment
  period to be simulated}

  \item{M.2}{the number of analyses during the follow-up
  period to be simulated}

  \item{alpha}{the error level at which the FDR is to be
  controlled}

  \item{powerThreshold}{the study is stopped as soon as the
  estimated power rate exceeds this threshold}

  \item{adjustment}{the method to use for the p-value
  adjustment to account for multiple testing.  In c("holm",
  "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
  "none")}

  \item{standalone}{boolean parameter that controls whether
  it is used within a bigger simulation.  Defaults to
  'TRUE'}
}
\value{
  In case the function is used in standalone mode:
  \item{resultTable}{the only real result. This table holds
  the error and power rates that came from the simulation.
  The fist row additionally shows at which analysis the
  simulated survival study was stopped.} The other values
  are copied versions of the corresponding input
  parameters.  They are included into the result for
  convenience only.

  In case the function is not used in standalone mode: The
  return value is one vector containing the elements of
  \code{resultTable}.
}
\description{
  interimTrial comes in two variants: When
  \code{standalone} is 'TRUE' the result is nicely
  formatted and can be used with the plotting functions
  \code{plotinterimFDR}, \code{plotinterimAPR}, and
  \code{plotinterimStops}.  When \code{standalone} is
  'FALSE' the result is presented as one single vector,
  which makes it usable within
  \code{interimTrialSimulation}.
}
\details{
  The \code{adjustment} parameter is used as parameter to
  \code{p.adjust}.

  The parameters \code{patdat} and \code{gendat} are of
  type \code{list} and do not necessarily have to be
  obtained from the functions \code{generatePatientData}
  and \code{generateExpressionData}.  Especially the
  entries \code{N.1} and \code{N.2} in \code{patdat} can be
  changed.  But in this case make sure to generate
  \code{gendat} accordingly.

  Internally interimTrial performs a gene-wise
  Cox-regression to determin the survival correlated genes.
}
\examples{
N <- 50
l.1.tick <- 60
lambda <- 60

patdat <- generatePatientData(N, l.1.tick, lambda)

fc <- c(rep(0, 500), ceiling(rnorm(500, 0, 1)-0.5))
Sigma.1 <- diag(1000)
Sigma.2 <- diag(1000)
N.1 <- patdat$N.1
N.2 <- patdat$N.2

gendat <- generateExpressionData(fc, Sigma.1, Sigma.2, N.1, N.2)

l.2 <- 60
M.1 <- 2
M.2 <- 2
alpha <- 0.05
powerThreshold <- 0.8
adjustment <- "BH"

\dontrun{T <- interimTrial(patdat, gendat, l.2, M.1, M.2, alpha, powerThreshold, adjustment)}
}
\author{
  Andreas Leha \email{andreas.leha@med.uni-goettingen.de}
}
\seealso{
  \code{\link{generatePatientData}} and
  \code{\link{generateExpressionData}} to generate the
  input data

  \code{\link{plotinterimFDR}},
  \code{\link{plotinterimAPR}}, and
  \code{\link{plotinterimStops}} to visualize the results

  \code{\link{interimTrialSimulation}} for a wrapper that
  simulates not one but many survival studies (and calls
  this function)
}

